+++ buildpacks/jvm-function-invoker/build.sh
@@ -4,5 +4,12 @@ set -euo pipefail
 buildpack_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 
 pushd "${buildpack_dir}"
-cargo make pack --profile "production"
+
+cargo libcnb package --release
+
+mv target target-cargo
+mkdir target
+cp -r target-cargo/buildpack/release/heroku_jvm-function-invoker/* target/
+cp package.toml target/
+
 popd